requirements.txt
cycler==0.10.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.8
kiwisolver==1.0.1
matplotlib==3.0.2
more-itertools==5.0.0
numpy==1.16.1
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.7.5
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
